requirement.txt
Django>=3.2,<4.0
psycopg2-binary>=2.8
djangorestframework>=3.12
djangorestframework-simplejwt>=4.8
Pillow>=8.2  # Required for image handling in Django
django-cors-headers>=3.7  # Required for handling CORS in the frontend
